honksquad: refactor: lock down PlayerBalanceComponent with [Access]
PlayerBalanceComponent was in Content.Shared but had no client or shared
consumers; only the server-side Economy systems used it. Moved it to
Content.Server/RussStation/Economy/ so [Access] can reference the owning
server system, and added [Access(typeof(PlayerBalanceSystem), typeof(PayrollSystem))].

BalanceCartridgeSystem now toggles mute through a SetPaycheckMuted setter;
tests seed Balance/JobId through SetBalance/SetJobId helpers on
PlayerBalanceSystem.

Part of #602.
